# swagger/contract-service-openapi.yaml
openapi: 3.0.3
info:
  title: Contract Service API
  description: |
    Contract Service for Mass Car Ordering Platform
    
    This service generates, stores, and retrieves official contract documents for deals 
    processed by the Purchase Request Service. It provides RESTful APIs to handle all 
    contract-related operations including PDF generation and event publishing.
  version: 1.0.0
  contact:
    name: Mercedes-Benz Contract Service Team
    email: contract-service@mercedes.com

servers:
  - url: http://localhost:8084/api/contract
    description: Local development server
  - url: http://contract-service:8084/api/contract
    description: Docker containerized server
  - url: https://dev-contract-service.mercedes.com/api/contract
    description: Development environment
  - url: https://staging-contract-service.mercedes.com/api/contract
    description: Staging environment
  - url: https://contract-service.mercedes.com/api/contract
    description: Production environment

paths:
  /v1/contracts:
    post:
      tags:
        - Contract Management
      summary: Generate a new contract
      description: |
        Creates a new contract from purchase request data, generates a PDF document,
        stores it in the configured location, and publishes a CONTRACT_CREATED event.
        
        **Business Flow:**
        1. Validates the incoming request payload
        2. Generates a unique contract ID
        3. Persists contract data to the database
        4. Generates PDF document using predefined template
        5. Stores PDF in configured location (local or S3)
        6. Publishes CONTRACT_CREATED event to Kafka
        7. Returns contract details with location
      operationId: generateContract
      parameters:
        - name: X-Trace-Id
          in: header
          description: Request correlation ID for tracing
          required: false
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractRequest'
            examples:
              hotel_fleet_example:
                summary: Hotel Fleet Contract Request
                value:
                  purchaseRequestId: "PR_12345"
                  dealId: "DEAL_ID_12345"
                  dealData:
                    dealId: "DEAL-5770C8DF"
                    customer:
                      customerAddress: "Leopoldstraße 45, 80802 München, Germany"
                      customerType: "HOTEL_FLEET"
                      customerPhone: "+49-89-123-4567"
                      customerEmail: "fleet@hilton.de"
                      preferredCurrency: "EUR"
                      customerId: "CUST-2025-SIMPLE"
                      customerCompany: "Hilton Hotels Deutschland"
                      customerName: "Hilton Hotels Deutschland GmbH"
                      customerTaxId: "DE987654321"
                    customerFinanceDetails:
                      type: "Lease"
                      provider: "Mercedes-Benz Financial"
                      approvalStatus: "Approved"
                      referenceNumber: "REF-12345"
                      termsInMonths: 36
                      interestRate: 0.045
                      downPaymentRequired: false
                    retailerInfo:
                      name: "Mercedes-Benz Downtown"
                      retailerId: "RTL-001"
                    massOrders:
                      - quantity: 1
                        massOrderId: "uuid-string"
                        priceBreakdown:
                          baseMsrp: 55000
                          optionsMsrp: 4300
                          totalMsrp: 59300
                          discountAmount: 4400
                          finalPrice: 54900
                        vehicleConfiguration:
                          brand: "Mercedes-Benz"
                          modelCode: "C300"
                          bodyType: "Sedan"
                          fuelType: "Gasoline"
                          transmission: "9G-TRONIC"
      responses:
        '201':
          description: Contract created successfully
          headers:
            Location:
              description: URI of the newly created contract
              schema:
                type: string
                format: uri
                example: "/v1/contracts/CONTRACT-0230EE56"
            X-Trace-Id:
              description: Request correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
              example:
                contractId: "CONTRACT-0230EE56"
                contractUrl: "/tmp/contracts/contract-0230ee56.pdf"
                contractStatus: "SIGNED"
                signedAt: "2025-10-16T11:13:16.161223Z"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "VALIDATION_FAILED"
                message: "Invalid request: Purchase request ID is required"
                timestamp: "2025-10-16T11:13:16.161223Z"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '409':
          description: Contract already exists for this purchase request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "CONTRACT_GENERATION_FAILED"
                message: "Contract already exists for this purchase request"
                timestamp: "2025-10-16T11:13:16.161223Z"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                timestamp: "2025-10-16T11:13:16.161223Z"
                traceId: "550e8400-e29b-41d4-a716-446655440000"

  /v1/contracts/{contractId}:
    get:
      tags:
        - Contract Management
      summary: Retrieve contract details
      description: |
        Gets the complete contract object in JSON format by contract ID.
        Returns all contract metadata including customer details, finance information,
        and vehicle orders as stored in the database.
      operationId: getContractById
      parameters:
        - name: contractId
          in: path
          description: Unique contract identifier
          required: true
          schema:
            type: string
            pattern: '^CONTRACT-[A-Z0-9]{8}$'
            example: "CONTRACT-0230EE56"
        - name: X-Trace-Id
          in: header
          description: Request correlation ID for tracing
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contract details retrieved successfully
          headers:
            X-Trace-Id:
              description: Request correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDetailsResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "CONTRACT_NOT_FOUND"
                message: "Contract not found"
                timestamp: "2025-10-16T11:13:16.161223Z"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/contracts/{contractId}/pdf:
    get:
      tags:
        - Contract Management
      summary: Download contract PDF
      description: |
        Downloads the PDF file for a specific contract. The PDF is generated during
        contract creation and stored in the configured location (local filesystem or S3).
        Returns the PDF file with appropriate Content-Type headers.
      operationId: downloadContractPdf
      parameters:
        - name: contractId
          in: path
          description: Unique contract identifier
          required: true
          schema:
            type: string
            pattern: '^CONTRACT-[A-Z0-9]{8}$'
            example: "CONTRACT-0230EE56"
        - name: X-Trace-Id
          in: header
          description: Request correlation ID for tracing
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF file downloaded successfully
          headers:
            Content-Type:
              description: MIME type of the response
              schema:
                type: string
                example: "application/pdf"
            Content-Disposition:
              description: Indicates file should be downloaded
              schema:
                type: string
                example: "attachment; filename=CONTRACT-0230EE56.pdf"
            X-Trace-Id:
              description: Request correlation ID
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
                description: PDF file content
        '404':
          description: Contract or PDF not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "CONTRACT_NOT_FOUND"
                message: "Contract not found"
                timestamp: "2025-10-16T11:13:16.161223Z"
                traceId: "550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check Endpoints
  /health/ready:
    get:
      tags:
        - Health Checks
      summary: Readiness probe
      description: |
        Checks if the service is ready to accept traffic.
        Validates database connectivity and other critical dependencies.
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "UP"
                  database:
                    type: string
                    example: "UP"
                  message:
                    type: string
                    example: "Service is ready"
        '503':
          description: Service is not ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "DOWN"
                  database:
                    type: string
                    example: "DOWN"
                  message:
                    type: string
                    example: "Database connection failed"

  /health/live:
    get:
      tags:
        - Health Checks
      summary: Liveness probe
      description: |
        Checks if the service is alive and responding.
        Used by container orchestrators to determine if the service should be restarted.
      operationId: livenessCheck
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "UP"
                  message:
                    type: string
                    example: "Service is alive"
                  timestamp:
                    type: integer
                    format: int64
                    example: 1697454796161

  /v1/contract/health:
    get:
      tags:
        - Health Checks
      summary: Business capability health check
      description: |
        Comprehensive health check for the contract service capability.
        Includes service status, database connectivity, and version information.
      operationId: contractHealthCheck
      responses:
        '200':
          description: Contract service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: "contract-service"
                  status:
                    type: string
                    example: "UP"
                  database:
                    type: string
                    example: "UP"
                  capability:
                    type: string
                    example: "contract"
                  version:
                    type: string
                    example: "1.0.0"
                  message:
                    type: string
                    example: "Contract service is healthy"
                  timestamp:
                    type: integer
                    format: int64
                    example: 1697454796161
        '503':
          description: Contract service is unhealthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: "contract-service"
                  status:
                    type: string
                    example: "DOWN"
                  database:
                    type: string
                    example: "DOWN"
                  capability:
                    type: string
                    example: "contract"
                  message:
                    type: string
                    example: "Contract service is unhealthy"
                  error:
                    type: string
                    example: "Database connection failed"

components:
  schemas:
    ContractRequest:
      type: object
      required:
        - purchaseRequestId
        - dealId
        - dealData
      properties:
        purchaseRequestId:
          type: string
          maxLength: 100
          description: Unique identifier for the purchase request
          example: "PR_12345"
        dealId:
          type: string
          maxLength: 100
          description: Unique identifier for the deal
          example: "DEAL_ID_12345"
        dealData:
          $ref: '#/components/schemas/DealData'

    DealData:
      type: object
      required:
        - dealId
        - customer
        - customerFinanceDetails
        - retailerInfo
        - massOrders
      properties:
        dealId:
          type: string
          description: Deal identifier within deal data
          example: "DEAL-5770C8DF"
        customer:
          type: object
          description: Complete customer information
          additionalProperties: true
          example:
            customerAddress: "Leopoldstraße 45, 80802 München, Germany"
            customerType: "HOTEL_FLEET"
            customerPhone: "+49-89-123-4567"
            customerEmail: "fleet@hilton.de"
            preferredCurrency: "EUR"
            customerId: "CUST-2025-SIMPLE"
            customerCompany: "Hilton Hotels Deutschland"
            customerName: "Hilton Hotels Deutschland GmbH"
            customerTaxId: "DE987654321"
        customerFinanceDetails:
          type: object
          description: Customer finance information
          additionalProperties: true
          example:
            type: "Lease"
            provider: "Mercedes-Benz Financial"
            approvalStatus: "Approved"
            referenceNumber: "REF-12345"
            termsInMonths: 36
            interestRate: 0.045
            downPaymentRequired: false
        retailerInfo:
          type: object
          description: Retailer information
          additionalProperties: true
          example:
            name: "Mercedes-Benz Downtown"
            retailerId: "RTL-001"
        massOrders:
          type: object
          description: Mass orders array with vehicle configurations
          additionalProperties: true
          example:
            - quantity: 1
              massOrderId: "uuid-string"
              priceBreakdown:
                baseMsrp: 55000
                optionsMsrp: 4300
                totalMsrp: 59300
                discountAmount: 4400
                finalPrice: 54900
              vehicleConfiguration:
                brand: "Mercedes-Benz"
                modelCode: "C300"
                bodyType: "Sedan"
                fuelType: "Gasoline"
                transmission: "9G-TRONIC"

    ContractResponse:
      type: object
      required:
        - contractId
        - contractUrl
        - contractStatus
        - signedAt
      properties:
        contractId:
          type: string
          pattern: '^CONTRACT-[A-Z0-9]{8}$'
          description: Unique contract identifier
          example: "CONTRACT-0230EE56"
        contractUrl:
          type: string
          maxLength: 500
          description: URL or path to the contract PDF
          example: "/tmp/contracts/contract-0230ee56.pdf"
        contractStatus:
          type: string
          enum: ["SIGNED", "PENDING", "CANCELLED"]
          description: Current status of the contract
          example: "SIGNED"
        signedAt:
          type: string
          format: date-time
          description: Timestamp when the contract was signed
          example: "2025-10-16T11:13:16.161223Z"

    ContractDetailsResponse:
      type: object
      required:
        - contractId
        - purchaseRequestId
        - dealId
        - customerDetails
        - financeDetails
        - massOrders
        - createdAt
      properties:
        contractId:
          type: string
          pattern: '^CONTRACT-[A-Z0-9]{8}$'
          description: Unique contract identifier
          example: "CONTRACT-0230EE56"
        purchaseRequestId:
          type: string
          maxLength: 100
          description: Reference to the originating purchase request
          example: "PR_12345"
        dealId:
          type: string
          maxLength: 100
          description: Reference to the deal
          example: "DEAL_ID_12345"
        customerDetails:
          type: object
          description: Complete customer information stored as JSONB
          additionalProperties: true
        financeDetails:
          type: object
          description: Customer finance details stored as JSONB
          additionalProperties: true
        massOrders:
          type: object
          description: Mass orders with vehicle configurations stored as JSONB
          additionalProperties: true
        pdfStorageLocation:
          type: string
          maxLength: 500
          description: File path or S3 URI for the generated PDF
          example: "/tmp/contracts/contract-0230ee56.pdf"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the contract was created
          example: "2025-10-16T11:13:16.161223Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the contract was last updated
          example: "2025-10-16T11:13:16.161223Z"

    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: Standardized error code for the error type
          enum:
            - "CONTRACT_NOT_FOUND"
            - "CONTRACT_GENERATION_FAILED"
            - "PDF_GENERATION_FAILED"
            - "VALIDATION_FAILED"
            - "INTERNAL_SERVER_ERROR"
          example: "CONTRACT_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "Contract not found"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2025-10-16T11:13:16.161223Z"
        traceId:
          type: string
          description: Request correlation ID for tracing (automatically captured from MDC)
          example: "550e8400-e29b-41d4-a716-446655440000"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication (if required in future versions)

# Global security (currently not implemented but defined for future use)
# security:
#   - BearerAuth: []

tags:
  - name: Contract Management
    description: Operations for contract generation and retrieval
  - name: Health Checks
    description: Health check endpoints for monitoring and orchestration